// runs/gemini-3-7-flash-high/stage1/src/model/hardware.ts
export interface HardwareState {
  // Performance
  master_level: number;
  pitch_stick: number;
  mod_wheel: number;

  // Organ
  organ_on: boolean;
  organ_model: number; // 0..5
  organ_preset_1: boolean;
  organ_preset_2: boolean;
  organ_db_16: number;
  organ_db_5_1_3: number;
  organ_db_8: number;
  organ_db_4: number;
  organ_db_2_2_3: number;
  organ_db_2: number;
  organ_db_1_3_5: number;
  organ_db_1_1_3: number;
  organ_db_1: number;
  organ_percussion_on: boolean;
  organ_percussion_soft: boolean;
  organ_percussion_fast: boolean;
  organ_percussion_third: boolean;
  organ_vibrato_on: boolean;
  organ_vibrato_mode: number;
  organ_rotary_stop: boolean;
  organ_rotary_speed: boolean;
  organ_octave_shift: number;
  organ_sustain: boolean;
  organ_pstick: boolean;

  // Piano
  piano_on: boolean;
  piano_layer_a_on: boolean;
  piano_layer_a_focus: boolean;
  piano_layer_a_level: number;
  piano_layer_a_octave: number;
  piano_layer_b_on: boolean;
  piano_layer_b_focus: boolean;
  piano_layer_b_level: number;
  piano_layer_b_octave: number;
  piano_type: number; // 0..5: Grand, Upright, Electric, Clav, Digital, Misc
  piano_model: number; // 1..9
  piano_kb_touch: number; // 0..3
  piano_timbre: number; // 0..5
  piano_dyn_comp: number; // 0..3
  piano_unison: number; // 0..3
  piano_soft_release: boolean;
  piano_string_res: boolean;
  piano_sustain: boolean;
  piano_pstick: boolean;

  // Program & Morph
  program_number: number; // 1..32
  program_button: number; // 1..8
  program_page: number;
  live_mode: boolean;
  layer_scene: number; // 1..2
  store: boolean;
  split: boolean;
  morph_wheel: boolean;
  morph_aftertouch: boolean;
  morph_ctrlped: boolean;
  transpose: boolean;
  panic: boolean;

  // Synth
  synth_on: boolean;
  synth_layer_a_on: boolean;
  synth_layer_a_level: number;
  synth_layer_b_on: boolean;
  synth_layer_b_level: number;
  synth_layer_c_on: boolean;
  synth_layer_c_level: number;
  synth_osc_type: number; // 0..3
  synth_osc_mod: number;
  synth_filter_type: number; // 0..4
  synth_filter_cutoff: number;
  synth_filter_resonance: number;
  synth_filter_drive: number;
  synth_filter_env_amt: number;
  synth_amp_attack: number;
  synth_amp_decay: number;
  synth_amp_sustain: number;
  synth_amp_release: number;
  synth_mod_attack: number;
  synth_mod_decay: number;
  synth_mod_release: number;
  synth_lfo_rate: number;
  synth_lfo_amount: number;
  synth_arp_run: boolean;
  synth_arp_rate: number;
  synth_unison: boolean;
  synth_vibrato: boolean;
  synth_sustain: boolean;
  synth_pstick: boolean;

  // Layer Effects
  effect_1_on: boolean;
  effect_1_type: number;
  effect_1_rate: number;
  effect_1_amount: number;
  effect_2_on: boolean;
  effect_2_type: number;
  effect_2_rate: number;
  effect_2_amount: number;
  delay_on: boolean;
  delay_tempo: number;
  delay_feedback: number;
  delay_amount: number;
  delay_pingpong: boolean;
  amp_eq_on: boolean;
  amp_type: number;
  amp_drive: number;
  eq_bass: number;
  eq_mid: number;
  eq_mid_freq: number;
  eq_treble: number;
  compressor_on: boolean;
  compressor_amount: number;
  compressor_fast: boolean;
  reverb_on: boolean;
  reverb_type: number;
  reverb_decay: number;
  reverb_amount: number;
  reverb_bright: boolean;
  layer_focus_piano: boolean;
  layer_focus_organ: boolean;
  layer_focus_synth: boolean;
}